Fix the remaining unit tests for 1.16.

+++ world/source/unit_tests/Weight_test.cpp
@@ -5,8 +5,8 @@ TEST(SW_World_Weight, constructors)
   Weight w(10, 10);
   Weight w2(1672);
 
-  EXPECT_EQ(170, w.get_weight());
-  EXPECT_EQ(1672, w2.get_weight());
+  EXPECT_EQ(static_cast<uint>(170), w.get_weight());
+  EXPECT_EQ(static_cast<uint>(1672), w2.get_weight());
 }
 
 TEST(SW_World_Weight, operator_eq)
@@ -67,19 +67,19 @@ TEST(SW_World_Weight, operator_leq)
 
 TEST(SW_World_Weight, get_lbs)
 {
-  EXPECT_EQ(0, Weight::get_lbs(0));
-  EXPECT_EQ(0, Weight::get_lbs(12));
-  EXPECT_EQ(1, Weight::get_lbs(16));
-  EXPECT_EQ(2, Weight::get_lbs(40));
-  EXPECT_EQ(100, Weight::get_lbs(1600));
+  EXPECT_EQ(static_cast<uint>(0), Weight::get_lbs(0));
+  EXPECT_EQ(static_cast<uint>(0), Weight::get_lbs(12));
+  EXPECT_EQ(static_cast<uint>(1), Weight::get_lbs(16));
+  EXPECT_EQ(static_cast<uint>(2), Weight::get_lbs(40));
+  EXPECT_EQ(static_cast<uint>(100), Weight::get_lbs(1600));
 }
 
 TEST(SW_World_Weight, get_oz)
 {
-  EXPECT_EQ(0, Weight::get_oz(0));
-  EXPECT_EQ(16, Weight::get_oz(1));
-  EXPECT_EQ(32, Weight::get_oz(2));
-  EXPECT_EQ(1600, Weight::get_oz(100));
+  EXPECT_EQ(static_cast<uint>(0), Weight::get_oz(0));
+  EXPECT_EQ(static_cast<uint>(16), Weight::get_oz(1));
+  EXPECT_EQ(static_cast<uint>(32), Weight::get_oz(2));
+  EXPECT_EQ(static_cast<uint>(1600), Weight::get_oz(100));
 }
 
 TEST(SW_World_Weight, serialization_id)
